Advance rip register when taking a snapshot
When taking a snapshot, the rip register is pointing to the instruction
that executed `outl`. If we restore to that point we will execute the
instruction again.

In the specific case of trying to take a SnapFaaS diff spanshot, the
result is _actually_ taking a nearly empty diff snapshot without
initializing the app whatsoever! Why? Because we load into the runtime
snapshot which, because the RIP register is pointing to the `outl`
instruction **immiediately takes the diff snapshot without doing
anything else!**.

The fix is to manually increment the RIP register (as the CPU would
do under regular execution) before serializing the captures registers to
the snapshot. Anecdotally, this pretty uniformly makes boot time for
_all_ applications around 5ms, several orders of magnitude decrease from
some of the heavier weight applications that require significant
initialization.